requirements.txt
tensorflow==2.16.1
pandas==2.2.1
dvc==3.49.0
mlflow==2.11.3
notebook==7.1.2
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
python-box[all]~=7.1.1
pyyaml==6.0.1
tqdm==4.66.2
ensure==1.0.4
joblib==1.3.2
types-pyyaml==6.0.12.20240311
scipy==1.12.0
flask==3.0.2
flask-cors==4.0.0
gdown==5.1.0
-e .